renaming membership_ui to access_card and fixing packagename of module
