+++ access_card/proguard-rules.pro
