+++ access_card/src/main/res/layout/fragment_validate_member_information.xml
