+++ access_card/src/main/res/values-en/strings.xml
