+++ access_card/src/main/res/values-es/strings.xml
