+++ access_card/src/main/res/values-zh/strings.xml
