+++ access_card/src/main/res/values/dimens.xml
