+++ access_card/src/main/res/values/strings.xml
