+++ access_card/src/main/res/values/styles.xml
